+++ src/rogallo/commands/navigation.py
@@ -21,4 +21,18 @@ class Backward(Command):
     SHOW_IN_FOOTER = True
 
 
+##############################################################################
+class GoToParent(Command):
+    """Go to the parent directory"""
+
+    BINDING_KEY = "alt+up"
+
+
+##############################################################################
+class GoToRoot(Command):
+    """Go to the root directory"""
+
+    BINDING_KEY = "alt+shift+up"
+
+
 ### navigation.py ends here