+++ src/rogallo/providers/main.py
@@ -21,6 +21,8 @@ from ..commands import (
     CopyLocationToClipboard,
     Forward,
     GoHome,
+    GoToParent,
+    GoToRoot,
     JumpToCommandLine,
     JumpToDocument,
     JumpToSidebar,
@@ -56,6 +58,8 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(CopyLocationToClipboard)
         yield from self.maybe(Forward)
         yield from self.maybe(GoHome)
+        yield from self.maybe(GoToParent)
+        yield from self.maybe(GoToRoot)
         yield Help()
         yield from self.maybe(JumpToCommandLine)
         yield from self.maybe(JumpToDocument)